perufauna: Easily Load and Harmonize Peruvian Fauna and Biodiversity Packages

The perufauna metapackage streamlines the installation, loading, and integration of specialized R packages focused on Peruvian fauna and biodiversity:

avesperu: Taxonomic backbone and search tools for the birds of Peru (UNOP checklist).

perumammals: Taxonomic backbone and validation for mammals of Peru (Pacheco et al. 2021).

citesperu: Status and verification of species in CITES Peru appendices (MINAM).

perufaunads004: Threatened fauna categorization according to D.S. N° 004-2014-MINAGRI and SERFOR Red Book.

In addition to orchestrated attachment, perufauna provides integrated cross-referencing functions such as pf_match and pf_status to analyze species across all taxonomic and regulatory sources in a single call.

See also

Author

Maintainer: Paul E. Santos Andrade paulefrens@gmail.com (ORCID) [copyright holder]

Authors:

Paul E. Santos Andrade paulefrens@gmail.com (ORCID) [copyright holder]